requirements/base.txt
# Base Requirement

cryptography==36.0.1
Django==4.0.3
django-extensions==3.1.5
djangorestframework==3.13.1
django-filter==21.1
django-phonenumber-field==6.1.0
djangorestframework-simplejwt==5.1.0
Pillow==9.0.1
psycopg2-binary==2.9.3
phonenumbers==8.12.42
pycrypto==2.6.1
requests==2.27.1
